Foi criado o modulo dashboard para organizar as features do dashboard do usuario logado
